Allow clicking of [+] to close per-object settings panel
Clicking the [+] button while the window is open now behaves the same way as the [x] button does. It detects whether the window is open by checking the window opacity. Checking base.currentIndex is unreliable since it doesn't get changed when you click the [x] or [+] button, only when you activate another object. Allowing the [+] button to be clicked while a different object's window is open would be even better, but may be an improvement for later.

Contributes to issue CURA-250.