Uncouple profile (un)serialisation from file saving/loading
In some cases it is desired to get a serialised version of a profile without having to save it to a file immediately. In those cases, you can now just call the serialise method of profiles. Unserialise is done in the same way to follow form.

Contributes to issue CURA-34.