Scale and mirror as separate operations
To prevent mirror from influencing the minimum scale and scale from being influenced by mirror, the scene node now keeps track of mirroring separately from scaling. This introduces a MirrorOperation which is different from ScaleOperation in that it changes a different variable in the scene node.

Contributes to issue CURA-612.