Convert tests to pytest framework
This framework allows for parameterised tests, meaning that all tests have to be rewritten to use a parameter that passes data to it for each test case that tests the same function. Currently, two tests still fail. One is because of a faulty function. The other I don't know why it fails (the test case should be the same!)

Contributes to issue CURA-260.