Round floats in QML rather than Python
In the display of the scale tool, the values are now rounded to 4 decimals. In the Python side they are no longer rounded. This fixes a floating point issue in Qt where it was displaying Javascript's floating point errors even after it was rounded in Python.

Contributes to issue CURA-607.