Fix error message when importing profiles
The error message wasn't displayed because it couldn't load the icon for the window. It was the wrong enum name. Also added a message for when it can't read the file at all, but don't worry: The internationalisation string is exactly the same as a string a few lines lower, so it won't add to the translation files.

Contributes to issue CURA-37.