Do not load a QtTexture until bind() is called
On 32-bit Windows, there is something going on with threads
that causes load() to be executed in a different thread from
the OpenGL Context. This then causes a crash. So instead, move
the texture loading to the next bind() call as we can be sure that
is called from the right thread.

Contributes to CURA-734