Use built-in wrapping functionality for tooltips
The trick is to put the text in a container that automatically adjusts to the size of the parent. The size of the parent is set by the operating system by how it judges tooltips to render, so this solution should behave more as the user expects it to behave.

Contributes to issue CURA-700.