Hook version upgrade in startup process
After loading the plug-ins, it will now do the version upgrades.

This means that the preferences have already been loaded at that point. They should then be changed (=todo)! This won't be possible for the language preference since that only works at a point in the start-up procedure before this hook. This means that the language might reset to default when the upgrade happens and the key for language is wrong.

Contributes to issue CURA-844.