Make pylint stop complaining about whitespace for keyword assignment
pylint insists that keyword assignment should not have spaces around the
"=". This is ugly and breaks with the rest of the code style, however we
cannot disable just that check and disabling bad-whitespace entirely
means we get no error reports about all the other issues. So instead,
make pylint stop complaining about keyword assignment by disabling it on
a per-line basis.