Moved load out of construction of setting registry
Because the load was in the constructor, which in turn calls the creation of container stacks (which require a container registry), we got an infinate loop.
By first creating the registry and then calling load, we ensure that __instance is set, so there is only one instance of it.

CURA-1278