Don't give a warning if hitting a file that was moved to /old
In a second pass it still has these file names, but some of them will have moved. Since the file lists are in a queue in no specific order, we can't really remove the files upon moving them to /old directory. Instead, we check for it here.

Contributes to issue CURA-844.