CMake: Remove hardcoded projectname
While updating Cura I noticed that translation files were still generated for Uranium (in a clean Cura build).
The reason for this is a hardcoded path, which needs to be replaced with CMake's PROJECT_NAME.

I remember there were discussions on JIRA about odd behaviour while checking our translations.
This might be one of the reasons...

Not sure which JIRA issue it really belongs to, but I set it as:
Contributes to CURA-570