Remove _open_args from SaveFile and use an explicit "encoding" parameter
Since we need to be sure to pass the right encoding to both
NamedTemporaryFile and open we need to know the encoding. We used to
handle this by specifying it in the _open_args but these were not used
by NamedTemporaryFile. Now we explicitly pass it to NamedTemporaryFile
and default to saving as utf-8.

Contributes to CURA-2692