Don't retry indefinitely for non-permission errors
It is understood that we must retry indefinitely for permission errors. That's not perfect for actual permission errors, but if the permission error is caused by a different application (or instance) keeping the file open, then that is the behaviour we seek. However for other exceptions, such as the file name being too long, we should let the error bubble up to the surface so that we can see it, or otherwise handle the error in the code that encounters it.

Contributes to issue CURA-2692.