Change config storage path to use back-slashes on Windows
Python magically fixes mixed slashes and backslashes in paths on Windows. However the storage path gets compared to the full path of files when loading configs to determine readonly state, and this comparison fails if the config storage path uses forward-slashes (resulting in all config files to be treated as readonly, including quality_changes and user profiles).