CURA-1764: Using set_tests_properties to add environment variables
As discussed on JIRA the best way to add our environment variables is to pass them via "set_tests_properties". Of course, this depends on #171, which begins to use the set of commands provided by CMake to declare tests.

However, there seems to be a problem in the test suite, therefore it isn't working completely now.

Depends on: #171
Contributes to CURA-1697
Contributes to CURA-1764