Print the full class path rather than the filename when logging
This avoids leaking build data file names into the logs. It also makes
it easier to see what logged the message since many of the names become
shorter. In addition, we avoid having a string of "..." in the logs
since we no longer need to replace parts of the log message.