UpdateChecker: Completing work and additions
As you can see from the diff the work on this plugin was not finished.
Moved the needed function into the job and added a message for the case
someone makes the job non-silent. The developer might think at the
moment, when seeing that the message appears: "No new version was
found." .. that everything is fine.
So to make it clear I added a message which is sent to the UIX.

Contributes to CURA-2970