Remove duplicate container_registry fixture in individual test files
This splits the container_registry fixture into container_registry and
loaded_container_registry, with the first one creating the registry and
the second one also loading it. It also makes sure the right fixture is
used in the individual test files.

This removes behaviour differences between running all tests and running
a single test.