Do not recreate the QQmlPropertyMap every time
This changes SettingPropertyProvider to not constantly recreate the
QQmlPropertyMap object, which reduces overhead when the provider is
updated and prevents Qt from getting confused with signal connections.

The only downside is that changing watched_properties may lead to stale
data in the QQmlPropertyMap. Since we do not change watched_properties
anywhere, I consider it safe to make this change.

Contributes to CURA-3312