Reinstate other search paths
The config path for Windows and OSX happens to be the same as their data and cache paths, but that's not the case for Linux. Also, on OSX it should also search the additional search paths in the resources in order to find the old location there. So it only worked properly on Windows.

Contributes to issue CURA-3479.