Don't search all Resources' search paths any more
It was once needed to allow updating Cura 2.1 to 2.2. However this also makes it search through all of the built-in search directories where the installation puts the profiles. This takes an unnecessary amount of time. This does prevent people from upgrading from 2.1 all the way to 2.5 however...

Contributes to issue CURA-3479.