Split used settings into used keys and used values
Used values actually need a value whereas used keys are only used to
indicate usage of settings (for dependency tracking). Everything
specified as a variable in Python expressions will add both a used value
and a used key, whereas strings will only add a used key.

This prevents recursion issues when dealing with "resolve" in Cura,
since the resolve for something like "material_bed_temperature" would
try to get the value of "material_bed_temperature" even when we are
trying to calculate the value of "material_bed_temperature".