Lock cache directory while working on it
The cache files are the last few bits that get written to the file system when Cura is closing. If you're restarting Cura while this is happening and it's loading the cache files while the previous Cura instance is still writing them, it might go wrong. This makes the new Cura instance wait for the old cache files to be written. Up to 10 seconds.

Contributes to issue CURA-4024.