Catch RecursionError when pickling definition for cache
We sometimes get a RecursionError, like 1 in 50 times, which causes the definition to not get added to the registry. Now we catch the error earlier so that the cache fails to save but we still properly add the definition. The cause of the RecursionError is still unknown.

Contributes to issue CURA-4024.