Removed depenency on webbrowser
We got reports that it's doing all kinds of weird stuff to figure out default browsers. It could well be that this is
what's causing the virus scanners to mark Cura as virus. Changing it to Qt function should not cause any issues,
as that's already used in a number of other places in the code