Rename LocalContainerProvider to LocalDefinitionProvider
We have a priority field in the plug-in metadata that determines the order in which the providers should be queried. We can use this instead of the current mechanism that hard-codes that definition containers have to be loaded first. All of the metadata loading and such will have to be different anyway.

Contributes to issue CURA-4243.