Implement getting metadata separately for every container
Due to inheritance, you need to be able to access the metadata of a different container that has not yet loaded. We'll do this lazily so each container only needs to really get loaded once. To allow that we need to adjust the interface of the ContainerProvider a bit.

Contributes to issue CURA-4243.