Set read_only of containers properly after loading
Previously a file was read-only if it was stored outside of the normal resource search directory for its resource type. This is no longer feasible if requesting separate IDs to load, because the resource type is unknown at that time. Instead we just look if the file is contained in the data directory. That is more semantically correct and makes the code much simpler.

Contributes to issue CURA-4243.