Add getIdsFromFile to containers
It has a default implementation that gets the ID from the file name by stripping the extension and unescaping it. This default implementation is currently in the 'interface', which is not neat, but it's not the only implementation in there so I guess it's all right.

Contributes to issue CURA-4243.