Make sure the metadata of a machine is loaded if you query for its ID
Normally it's unused, but during the loading of all metadata, some other metadata fields are requested as well, such as the inheritance fields for definitions. This allows lazy-loading during the loading of all metadata, so that the order in which the definitions are loaded doesn't matter any more.

Contributes to issue CURA-4243.