Keep track of a mix of complete containers and metadata
This makes the model implementation quite a lot more complex, but it's the only solution we could come up with. The model now contains two sets of containers: One of containers that are fully loaded and one of containers where only the metadata is known. When loading a container fully, that container is moved to the other dictionary.

Contributes to issue CURA-4243.