Make queries more efficient when they contain ID and something else
When they contained only an ID we'd just return the item from the dictionary. Otherwise we would perform a full query.

Now when they contain an ID we'd have a candidate and then check the rest of the metadata properties on that candidate. So now if we have a query that requests, for instance, the ID as well as the container type, then this would also not require a full container query.

Contributes to issue CURA-4243.