Fill default metadata if setting metadata
We sometimes had cases where the metadata was being set after the name, ID and definition are being set. These have all been fixed now, except in some tests. Because of these tests I decided that the setMetaData should be robust against this. You can override the ID, name or definition via the setMetaData function but if it's not overwriting them it will not erase these required fields.

Contributes to issue CURA-4243.