CURA-4935 Can't disable sending anonymous slice data and checking updates
This is the ~5 minute~ 30 second fix. I know it's not the more proper way to do it, but I've spent several hours on the more proper way to do it and it's causing huge amounts of headache (because, for example, the PluginBrowser installs plugins and the PluginRegistry removes them, and the PluginBrowser can't update the PluginsModel because the PluginsModel is intantiated in QML instead of in `PluginBrowser.py`, etc. etc. etc.

I'm effectively doing a full refactor of this functionality while simultaneously avoiding moving any "install from remote URL" functionality into the PluginRegistry which is dumb.

I'd prefer a simple, stable soution now instead of re-writing this all again.

I'll make a proposal for a proper, organized refactor for Cookies & Cura.