CURA-4968 Type-in transform does not go to Z=0 for centered scene nodes
Side effect: when a model is dropped on the build plate, the Z value displayed is now never 0, but instead 1/2 of its height. But that's where the object's center techncially is so IMO this is actually _more_ intuitive than the old system where the gizmo and type-in were at the bottom of the bounding box while X and Y were at the center of the bounding box.

Also, this still works with "automatically drop models on build plate" so yay!