Let removeContainer work correctly if container can't be loaded
So don't try to load the container in order to remove signals from it. The logic error here was that we would need to load the container in order to remove signals from it, but before the container was ever loaded you could never have added signals to that container anyway.

Contributes to issue CURA-5045.