Don't load metadata of plug-in if API version is incorrect
I had a plug-in with an incorrect API version, but the metadata still got loaded. As a result, my save-to-file dialogue still said I could write using the file type provided by my plug-in, but when I actually tried it could not find the writer. This prevents the metadata from appearing as well.