Don't shift containers if one of them didn't load
If a container couldn't load, previously the position of all containers in the stack would have shifted (quality in the place of material, quality_changes in the place of quality, etc). Now this places an empty container in there, which is not ideal either but at least it doesn't crash and the user gets an error message that something is up with their container stack.

Part of a fix for Cura issue #3570 on Github.