Hack: Add exception for OctoPrint plug-in
Earlier versions crash Cura. We can't re-add all of the old API with deprecated markers, so we're disabling this plug-in with a hard-coded version here now. Otherwise people who added a connection via OctoPrint aren't able to start Cura at all. Now they are able to start Cura and will find that OctoPrint doesn't work any more, and they should be able to update the plug-in.

Contributes to issue CURA-5287.