Remove all stubs
These have internal inconsistencies in them because they may be made on computers with a different PyQt version and different Mypy version than the ones on which they are run. The people who run this test should have the dependencies properly installed anyways so then the stubs are not necessary.

Contributes to issue CURA-5330.