Remove unused QML files
These have been specialised in Cura and are not used by any other projects using Uranium. The SettingVisibilityPage should be in Uranium in my opinion but that would involve moving the SettingVisibilityPreset class to Uranium too (as it should be, actually) which is a little more involved than I would dare to get into now.

Contributes to issue CURA-5330.