Don't raise false alarm if interface is instantiated
This would give a false alarm since the decorator checks if a method is the same method object as the method with the same name as the subclass. But if the subclass is the same then this would always give an error because no method has been overwritten from the interface (since it is the interface itself).

Contributes to issue CURA-5330.