Revert "Set container's parent to application by default"
This reverts commit 52b1e9b9e0a2ed9e5186075a2eaa514117aa4829.
That commit was just extremely dangerous. Putting the APPLICATION as the parent QObject of every container causes the application to be pickled along when you pickle a container, which will crash everywhere.
This commit was also not documented at all and not linked to any ticket. It seems completely unnecessary and a misunderstanding of how PyQt works.

Contributes to issue CURA-5330.