Fix broken iterator types
Seems like there is a bug in MyPy here. It's complaining that the BreadthFirstIterator doesn't have a __next__ function. And indeed it doesn't, but using the for x in y: syntax it should first call iter() on the BreadthFirstIterator and then you get an object that does have a __next__ method, so it should be all right.

Contributes to issue CURA-5330.