Don't allow Qt to delete the object
This indicates that we need 'C++' to take the ownership. In reality, that then becomes PyQt and in effect Python will take ownership. This is necessary because we need to pass None as the parent of this QObject, otherwise Pickle will pick it up and pickle it along and produce humongous pickle files, slowing the start-up of the application down a lot. With this flag we say that we want the ownership of the object ourselves so that Qt doesn't delete the C++ object. Normally it would do that if the parent is None.

Contributes to issue CURA-5330.