Limit sources to the version upgrades which defined them
CURA-5440

If a configuration file source directory is defined in an early version,
later versions will use that same directory. For example:

In version 1:
  user_config: src_dir="./all" dst_dir="./user" src_version=1.1 target=2.1
  app_config:  src_dir="./all" dst_dir="./app"  src_version=2.0 target=2.1

In version 2:
  user_config: src_dir="./user" dst_dir="./user" src_version=2.1 target=2.2
  app_config:  src_dir="./app"  dst_dir="./app"  src_version=2.1 target=3.0

So before version 1, both "user_config" and "app_config" can come from
"./all", but onwards, they are saved in "./user" and "./app"
respectively.

This commit makes sure that when we scan for and use app and user config
files in "./all", this should only happen when the app file is of version
1.1 and the user file is version 2.0.