Consistently return set from getAllKeys
ContainerStack and DefinitionContainer were returning sets, but InstanceContainer was casting the resulting set to a list first. Turns out that everywhere where we are using this, we are either casting it to a set or just iterating over it. So just don't cast it to a list please.

Contributes to issue CURA-5330.