Reduce drag distance
This makes the camera more responsive when you rotate the camera. It makes it slightly harder to access the context menu of the main window if you have Parkinson's disease, or just a very sensitive touch pad that makes you use a gesture to right click. But I think that was just way to wide of a drag distance and it's still pretty easy to right-click without dragging.

Some movement while right clicking is still allowed. About 10 pixels on my 1080p screen.