Clip to zoom when zooming
This prevents the camera from being stuck if the maximum zoom range changed while you were farther away than the new maximum zoom range. It would get stuck because the new r value in _zoomCamera would be out of bounds, so it would not execute the zoom. However this also prevents you from zooming farther in.