Move MIME type declaration into constructor of STLReader
Don't define new MIME types statically in a plug-in. Because the static code is executed upon importing the .py file, and the .py file is imported when reading the metadata of a plug-in. This metadata is read in order to judge what API version the plug-in has as well as to see whether it is disabled. So if you disable the plug-in, the MIME type would still be declared.