Define the name of the lock file in the application instead of in the Container Registry.
So now every application can define its own file name. This is also done
because we want to get rid of all uranium.lock files in the users'
folder, since we notice that there are many cases in which the file is
imposible to delete programatically in Windows due to permission errors.