Don't create list from intermediary filter results
This could cost a lot of performance. Originally it was keeping it as a filter because several wrapped filters will be executed lazily, but creating a list from it at every filter iteration forces it to filter everything every loop and allocate a list for it too.
This change was made in order to fix typing, so fix the typing as well, please.

Contributes to issue CURA-5796.