Don't require UraniumLibraryDir any more
I was getting annoyed by this LibraryDir.py file showing up in my Git. I could just have put it in .gitignore but I'm opting for a simpler solution: Just put all possible options on the plug-in location list. This list isn't evaluated that often, and when it is we first check whether the directory exists or not. So the ones that don't exist in these options will be ignored.