Added a direct copy to the matrix class as this prevents data conversion
The getData actually convers from float64 to float32. Changing this causes
issues with rendering, so there is probably one part of the code that really needs / wants the 32 version.

As I don't know which one that is, the best I can do is prevent having to do the conversion in some of the cases.