Fix reloading duplicated objects
If an object was duplicated and the file it came from was changed, it would show a message asking the user to reload the mesh. Clicking the reload button would only reload the last mesh though, because that's the one that self._callback was storing.

Now it reloads all of them.