Disable refchecks while resizing arrays
It only crashes the application if it occurs. Disabling this allows PyCharm to hook into these arrays to see their contents. Because of that hook, Numpy is tracking that and seeing that something could still be displaying the array and disallows resizing. If I disable the refcheck I can launch Cura in PyCharm's debugger.