Expose face-ID's via the selection-shader.
Instead of just passing the object-id, also show each gl_PrimitiveID in the shader output.
In order to make it still work for old, old OpenGL versions, I had to hide the information in gl_FragColor somehow: The upper bits of the color repressent the object, to lower bits repressent the faces.