UM.Controller: Adding variable for tool fallback
Need to clear all tools on my scene and add my own "toolset". However, in this toolset there is no TranslateTool and I would like to fall back to a different tool.
Of course, it would be convenient to have also get* and set* functions, but I don't care about these right now.