Don't include own type in query cache key
This is only relevant if we're subclassing these queries and change the results of the subclass with the same keys. This is not relevant right now. It could be relevant in the future if we want to subclass this class but right now it's hurting performance so it has to go.

Contributes to issue CURA-6793.