Cache partial queries as well
This is a huge performance increase. Previously, to get e.g. all material profiles for Ultimaker S5 with AA0.4 nozzles, we'd have to filter all profiles for material type, then for UMS5, then for 0.4. Then if we'd request all material profiles for UMS5 with AA0.8, we'd have to restart the filter for material type, then re-filter UMS5, and finally filter for AA0.8 instead of AA0.4.
With this change we'd store the query for all material types. And then store the query for UMS5, so for the second query we'd only need to go through the cached material profiles for UMS5 to filter for AA0.8 instead of AA0.4 and we'd prevent having to re-do 90% of the work.

Contributes to issue CURA-6793.