Fix clearing cache specific to a container
It is now no longer biased to the container type.
This bias was giving wrong results, since the container type was not always present in all queries. When it wasn't, it would give query results that contain certain containers but those query results weren't invalidated when a container got added because the query wasn't by container type. I think this has been a long standing problem but may just have not been surfacing by accident.

Contributes to issue CURA-6793.