Remove cache limit
We've found that this limit was never hit any more, so we'll just remove it. The memory impact of a cache entry like this is not so big anyway: It's just a few string pointers for the keys and a ContainerQuery instance as the value, which is in itself again a few pointers. It could grow to a couple of megabytes of memory, but nothing significant in comparison with the 3D model or CuraEngine.

Contributes to issue CURA-6793.