Inital work on 'Singing of Plugins and Material Packages'.
For the plugins: The scaffolding is built (that is, the structure planned to be used right now is in place) but the actual checksum-functionality yet needs to be built.
For the material packages: Nothing done yet, but should be easier than the plugins even, can reuse a lot of code probably.

part of CURA-6856