Move package verification check to Uranium containers.
Also make it so that these files are only _verified_ if there _is_ a signature, otherwise continue as normal.
This is opposed to what happens w.r.t. plugins, where if there isn't a signatures file, the process aborts and the plugin isn't loaded (it _has_ to be checked, the user can't just delete the check file).

part of CURA-6856