(For now) skip key-verification on public key not found.
Since we can't agree on the precise details of the signing process with the other teams, the Enterprise edition POC is probably going forward withouth the added security measures.
This commit adds a stopgap measure to not crash on the missing public key (which doesn't get added to Enterprise builds yet, because there isn't any official public --or private for that matter-- key yet).
Like I say in the comment: This may not be as bad as it seems, since, even if we forget to remove these lines, then, _as long as we remember to include a public key in the build_ it'll still work (also because a non-admin enterprise user won't have access to the location of the public key file).

part of CURA-7016