Re-use already made OpenGL 4.1 Context.
Some Intel drivers (though they may have fixed that bug this week) contain an issue where they might cause programs to hang if an OpenGL context is deleted. Normally this is at the end of programs so not a big issue, but for us we made and OpenGL 4.1 context, then separately made another OpenGL 4.1 context to check some things, then (implicitly by going out of scope) deleted the one that was only used for checking. This caused us to hang sometimes right on start-up for specific combinations of Windows/Intel/grpahics-card/driver/machine.

part of CURA-6670 and also hopefully of CURA-7013