Revert perspective changes to Camera.py in 64e88d51312eaafee406fedc4bbda547ea2e12c3
This way the perspective would change when zooming in if the camera is in perspective mode. Instead we use dollying, so the perspective shouldn't change.
This was also breaking some of the unit tests for Camera, which check hard-coded camera orientations that were sampled from when it was considered to be working correctly.

Contributes to issue CURA-6560.