Limit recursive JSONType to just 3 levels
This means that the function will give typing errors if you call it with dictionaries or lists that are more than 3 levels deep:
{
    key: {
        key: {
            key: {
                key: val
            }
        }
    }
}
This will give a typing error now.
This was necessary to prevent a crash of MyPy due to bug https://github.com/python/mypy/issues/731 .

Contributes to issue CURA-6627.