Fixes simulation sliders yet again.
If a mesh is drawn with a range, make sure the current range is applicable to that mesh. (But don't reset the range to the max of that mesh.)
What this does is fix the OpenGL crash that you can get when the max is over the index-buffer length -- but without messing up the sim. sliders.