Don't use Cura's i18n catalogue for this translation
Because self._i18n_catalog is not defined by QtApplication or Application. It's only defined in CuraApplication which is why this worked for Cura, but then it loaded the wrong i18n catalogue which doesn't have the translation for Loading UI... so it wouldn't find this translation.

Contributes to issue CURA-7201.