Fix loading meshes that contain texture coordinates but not normals
The format for a face is as follows:
  f v1/t1/n1 v2/t2/n2 v3/t3/n3
The part after a slash is optional. You could also have faces like this, without the normals:
  f v1/t1 v2/t2 v3/t3
Or even a face like this, with just the vertices and no textures:
  f v1 v2 v3
However it's not possible to have a face with vertices and normals but no textures:
  f v1/n1 v2/n2 v3/n3
This gets interpreted as a face with vertices and textures.

When given a face with just vertices and textures, our old code would give an IndexError saying that the normal vectors are out of range. That would be correct since it only stored -1 there. This fixes that by turning the texture and normal data around in our internal representation so that it has the same representation as the actual OBJ file. The same order. And so the len() checks if they are present will have the same effect as a len() check on the number of slashes in the original file.

Found during Turbo Testing and Tooling.