Prevent crashing when theme file couldn't be loaded for any reason
The EnvironmentError catches cases where the theme file is missing, can't be read due to permissions, etc. I haven't seen this in the field but while I'm here we might as well give a proper error message when this happens.
The UnicodeDecodeError has been seen in the field. People put weird font names in their themes.

Fixes Sentry issue CURA-NK.