Catch environment errors while saving package management file
Could be that the directory permissions changed just then, or got deleted while Cura is running. This way it won't crash (but will be unable to store the package management file). Chances are that it won't be able to log anything then either, but we should at least try to log it.