Fix crash if output devices get updated while a new device is added
The getOutputDevices() function returns a view on the values of the actual dictionary of output devices of the device manager. On separate threads (such as the zeroconf thread) it may add new output devices. If you're very unlucky, it may add a new output device during that microsecond while you're iterating over those devices. And of course, with our users being quite numerous, there's always some people that are actually that unlucky.

Modifying a sequence while iterating over it is not allowed. Instead, we make a local copy and then iterate over that.

Fixes Sentry issue CURA-11J.