Prevent ZeroDivisionError for applications that don't have a minimum window size
The mouse position must always be 0 if the window size is 0.
If the window has a width of 0 but not a height of 0, or the other way around, then this would unnecessarily also set the height to 0, but really the mouse position is then very weird anyway and checking for all of that would take more processing power than it's worth.

Fixes Sentry issue CURA-12F.