Prevent crash if cache folder can't be deleted
Apparently it can happen that a file doesn't exist any more but os.path.exists still returns true. This can happen because the operating system is slow to delete files (e.g. the subfiles have been registered by Python for deletion but are already deleted) or because of concurrency issues (a separate process already deleted the file in between the check if the file exists and the actual deletion).
In any case, I've made it try again because it's important that this cache gets deleted when first loading a new version of Cura. And if it still fails, it's probably a permission error or something and we can't really do anything about it. At least let's not crash. We'll put an error in the log file about it.

Fixes Sentry issue CURA-139.