Fix plugins not loading the second time
Once plugins load, the QML engine generates compiled qml (.qmlc) files to keep them cached.
Unfortunately, these files cannot be verified and signed as they are generated on runtime.
Thus, the second time the plugin loads, the signature check fails because it finds files that
are not considered "trusted".

To fix this issue, Trust now deletes those files when checking the plugin, to make sure that
their signature check does not fail.

This _may_ slow down the loading process of these plugins a tiny bit, but since the qmlc files
are still generated when the plugin is loaded or used, subsequent uses of the plugin in a session
should still be optimal (from a performance perspective).

Fixes CURA-7727