Correct MIME type for STL
It would've been x-stl-binary before (dash instead of period) which was causing an error and the files not to be saved since the printer doesn't support this MIME type then.
But there has been a standardised MIME type for STL since then: https://www.iana.org/assignments/media-types/model/stl . So we should use this one.