Add option to load certain plug-ins before the rest
In particular an application might want to load the logging plug-ins first. Currently Cura manually calls the loadPlugin function for ConsoleLogger, and then ConsoleLogger is also loaded again later which is causing warnings. This prevents that because Uranium knows which plug-ins are pre-loaded and will not attempt to load those again.
This is a partial solution. A complete solution would be to allow plug-ins to depend on other plug-ins. All plug-ins that log things would then need to depend on the logger plug-ins. That system is more complex though so I'll not implement that until it's deemed necessary.

Contributes to issue CURA-7501.