Fix warning of float face count when loading STLs with fallback
This is the fallback STL implementation, if numpy-stl can't be loaded (e.g. on Linux).
The number of faces was a float due to division by 3. We now round that down to an integer number of faces, just like the MeshBuilder does when it encounters a float. This prevents a warning in the log.
I've also adjusted the input type so that we don't just get a warning at runtime but also a warning in our static analysis.

Contributes to issue CURA-7501.